+++ pandera/api/pandas/__init__.py
@@ -1,0 +1,5 @@
+"""Pandas core."""
+
+from pandera.api.pandas.array import SeriesSchema
+from pandera.api.pandas.components import Column, Index, MultiIndex
+from pandera.api.pandas.container import DataFrameSchema
