+++ pandera/api/pandas/components.py
@@ -12,9 +12,9 @@ from pandera.backends.pandas.components import (
     IndexBackend,
     MultiIndexBackend,
 )
-from pandera.core.pandas.array import ArraySchema
-from pandera.core.pandas.container import DataFrameSchema
-from pandera.core.pandas.types import CheckList, PandasDtypeInputTypes
+from pandera.api.pandas.array import ArraySchema
+from pandera.api.pandas.container import DataFrameSchema
+from pandera.api.pandas.types import CheckList, PandasDtypeInputTypes
 from pandera.dtypes import UniqueSettings
