+++ pandera/api/pandas/container.py
@@ -13,10 +13,10 @@ import pandas as pd
 from pandera import errors
 from pandera import strategies as st
 from pandera.backends.pandas.container import DataFrameSchemaBackend
-from pandera.core.base.schema import BaseSchema, inferred_schema_guard
-from pandera.core.checks import Check
-from pandera.core.hypotheses import Hypothesis
-from pandera.core.pandas.types import (
+from pandera.api.base.schema import BaseSchema, inferred_schema_guard
+from pandera.api.checks import Check
+from pandera.api.hypotheses import Hypothesis
+from pandera.api.pandas.types import (
     CheckList,
     PandasDtypeInputTypes,
     StrictType,
@@ -723,7 +723,7 @@ class DataFrameSchema(BaseSchema):  # pylint: disable=too-many-public-methods
 
         """
         # pylint: disable=import-outside-toplevel,import-outside-toplevel
-        from pandera.core.pandas.components import Column
+        from pandera.api.pandas.components import Column
 
         new_schema = copy.deepcopy(self)
 
@@ -987,7 +987,7 @@ class DataFrameSchema(BaseSchema):  # pylint: disable=too-many-public-methods
 
         """
         # pylint: disable=import-outside-toplevel,cyclic-import
-        from pandera.core.pandas.components import Index, MultiIndex
+        from pandera.api.pandas.components import Index, MultiIndex
 
         new_schema = copy.deepcopy(self)
 
@@ -1114,7 +1114,7 @@ class DataFrameSchema(BaseSchema):  # pylint: disable=too-many-public-methods
 
         """
         # pylint: disable=import-outside-toplevel,cyclic-import
-        from pandera.core.pandas.components import Column, Index, MultiIndex
+        from pandera.api.pandas.components import Column, Index, MultiIndex
 
         # explcit check for an empty list
         if level == []:
