+++ pandera/backends/pandas/builtin_checks.py
@@ -7,7 +7,7 @@ from typing import cast, Any, Iterable, TypeVar, Union
 import pandas as pd
 
 import pandera.strategies as st
-from pandera.core.extensions import register_builtin_check
+from pandera.api.extensions import register_builtin_check
 
 from pandera.typing.modin import MODIN_INSTALLED
 from pandera.typing.pyspark import PYSPARK_INSTALLED
