--- pandera/core/pandas/__init__.py
@@ -1,5 +1,0 @@
-"""Pandas core."""
-
-from pandera.core.pandas.array import SeriesSchema
-from pandera.core.pandas.components import Column, Index, MultiIndex
-from pandera.core.pandas.container import DataFrameSchema
