+++ pandera/io/pandas_io.py
@@ -12,9 +12,9 @@ import pandas as pd
 import pandera.errors
 
 from pandera import dtypes
-from pandera.core.pandas.container import DataFrameSchema
-from pandera.core.pandas.components import Column
-from pandera.core.checks import Check
+from pandera.api.pandas.container import DataFrameSchema
+from pandera.api.pandas.components import Column
+from pandera.api.checks import Check
 from pandera.engines import pandas_engine
 from pandera.schema_statistics import get_dataframe_schema_statistics
