+++ tests/core/checks_fixtures.py
@@ -6,7 +6,7 @@ import pandas as pd
 import pytest
 
 import pandera as pa
-import pandera.core.extensions as pa_ext
+import pandera.api.extensions as pa_ext
 
 __all__ = "custom_check_teardown", "extra_registered_checks"
